fixing code climate issues
